a15b - Make the Creature class's getSpecies() function a pure virtual function.
Creature class's getSpecies() function never gets called. However, it
is absolutely critical that any class that is derived from the Creature
class define a getSpecies() function since that function is called from
the Creature class's getDamage() function. The best way to implement
this is to make the Creature class's getSpecies() function a pure
virtual function, so that every class that is derived from the Creature
class will be required to implement a getSpecies() function.